fix bugs in deploy scripts

+++ 08-deploy.sh
@@ -2,6 +2,11 @@
 # prepare a pre-snapshot and move it to Nyx-images
 target=$1
 
+if [[ ! -d ../Nyx-images/pre_snapshot ]]; then
+    echo "Please prepare a pre-snapshot in ../Nyx-images"
+    exit
+fi
+
 cd ../Nyx-images
 for idx in {0..9}; do
     rm -rf pre_snapshot-${target}-${idx}
